--- a/halo2_proofs/src/multicore.rs
+++ b/halo2_proofs/src/multicore.rs
@@ -2,4 +2,4 @@
 //! `halo2`. It's currently just a (very!) thin wrapper around [`rayon`] but may
 //! be extended in the future to allow for various parallelism strategies.
 
-pub use rayon::{current_num_threads, prelude, scope, Scope};
+pub use rayon::{current_num_threads, join, prelude, scope, Scope};